feat(table): added CRUD operations to the table
This also included support for reading from global indexes.
