+++ .golangci.yml
@@ -51,7 +51,7 @@ linters-settings:
           - (github.com/golangci/golangci-lint/pkg/logutils.Log).Errorf
           - (github.com/golangci/golangci-lint/pkg/logutils.Log).Fatalf
   lll:
-    line-length: 140
+    line-length: 200
   maligned:
     suggest-new: true
   misspell:
